Add CFG option to limit course browser to certain rights levels
Adds $CFG['coursebrowserRightsToPromote'] option as minimum rights to
be able to promote a course in the course browser. Also allows an admin
to promote another user's course or make it a template without losing that
setting when the user modifies the course settings.  Also fixes the
"contributed by" when not the user's course.