Bug fix: prevent overlapping ajax calls from assess2 frontend
This is meant to address race conditions by preventing a subsequent front end
call while one is still processing.

Also, in some browsers the autosave was happening even after submission, suggesting
a different race condition between the submitQuestion clearing the autosavetimer and
the blur adding an autosave timer.  The blur should happen first, but just in case,
this adds a slight delay to the clearautosave call to ensure the blur call has processed
first.
IM-758