Adjust getstuans to return null on blank stuanswers by default IM-878
In the new player, on the scoring pass all parts populate stuanswers, even if
they weren't submitted for scoring. But a lot of tests use !== null, so
this has getstuans to return null on a blank answer by default. This shouldn't
affect scoring of blank answers, which relies on the original stuanswers