I extended functionality of finderiv_payout to allow for array inputs and bad types. I added a function finderiv_checkpayout that compares a payout with lines returned from a chart.
The functionality for payout was to allow the final price to be either a single value or an array of final prices. If a single value then the function returns a single value. If an array then the function returns an array of values. Also, payout now ignores positions that do not have a valid type. This allows the user to input empty positions and works well with student answers.

THe function checkpayout simplifies checking a student answer where you have given them a portfolio and ask them to draw a picture of the payout diagram. The data is returned as a series of lines and you need to match the line to the payout.

The documentation has been updated for these changes.

The test question has been updated for these changes (606302).